+++ clients/web/react/src/RegisterPage/RegisterPage.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect, useRef } from 'react';
-import { Link, useHistory } from 'react-router-dom';
+import { Link, useNavigate } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
 import { Button, Row, Col } from 'react-bootstrap';
 
@@ -9,14 +9,14 @@ import { ServerVerifiedPin } from '../_components';
 import { create, supported } from '@github/webauthn-json';
 import base64url from 'base64url';
 import cbor from 'cbor';
-import { Auth } from 'aws-amplify';
+import { signUp, signIn, confirmSignIn, fetchAuthSession } from 'aws-amplify/auth';
 
 function RegisterPage() {
 
     const username = localStorage.getItem('username');
     const [submitted, setSubmitted] = useState(false);
     const signInResult = useSelector(state => state.authentication.signInResult);
-    const history = useHistory();
+    const navigate = useNavigate();
     const [cognitoUser, setCognitoUser] = useState({});
     const defaultInvalidPIN = -1;
     const dispatch = useDispatch();
@@ -27,7 +27,7 @@ function RegisterPage() {
         dispatch(userActions.logout());
         console.log("signInResult: ", signInResult);
         return (() => {
-            if (history.action === "POP") {
+            // Note: history.action check removed - not available in React Router v6
                 // Code here will run when back button fires. Note that it's after the `return` for useEffect's callback; code before the return will fire after the page mounts, code after when it is about to unmount.
                 dispatch(credentialActions.completeUV());
             }
@@ -53,10 +53,12 @@ function RegisterPage() {
         const attributes =  {"name": usernameLower};
 
         try {
-            const { user } = await Auth.signUp({
+            const { user } = await signUp({
                 username,
                 password,
-                attributes
+                options: {
+                    userAttributes: attributes
+                }
             });
             console.log("SignUp: ", user);
         } catch (error) {
@@ -66,13 +68,13 @@ function RegisterPage() {
 
         try {
 
-            let cognitoUser = await Auth.signIn(username);
+            let cognitoUser = await signIn({ username });
             setCognitoUser(cognitoUser);
             console.log("SignIn CognitoUser: ", cognitoUser);
 
             if(cognitoUser.challengeName === 'CUSTOM_CHALLENGE' && cognitoUser.challengeParam.type === 'webauthn.get'){
                 dispatch(alertActions.error("You have already registered. Please sign in."));
-                history.push('/login');
+                navigate('/login');
                 return;
             }
 
@@ -111,23 +113,23 @@ function RegisterPage() {
                 if(uv == false) {
                     dispatch(credentialActions.getUV(challengeResponse));
                 } else {
-                    console.log("sendCustomChallengeAnswer: ", cognitoUser);
+                    console.log("confirmSignIn: ", cognitoUser);
                     // to send the answer of the custom challenge
-                    const user = await Auth.sendCustomChallengeAnswer(cognitoUser, JSON.stringify(challengeResponse))
+                    const user = await confirmSignIn({ challengeResponse: JSON.stringify(challengeResponse) })
                     .then(user => {
                         console.log(user);
-                        
-                        Auth.currentSession()
-                        .then(data => {
+
+                        fetchAuthSession()
+                        .then(session => {
                             dispatch(alertActions.success('Registration successful'));
                             let userData = {
                                 "id": 1,
-                                "username": user.attributes.name,
-                                "token": data.getAccessToken().getJwtToken()
+                                "username": user.username || user.signInDetails?.loginId,
+                                "token": session.tokens?.accessToken?.toString()
                             }
                             localStorage.setItem('user', JSON.stringify(userData));
                             console.log("userData ", localStorage.getItem('user'));
-                            history.push('/');
+                            navigate('/');
                         })
                         .catch(err => {
                             console.log("currentSession error: ", err);
@@ -226,7 +228,7 @@ function RegisterPage() {
                     }
                     localStorage.setItem('user', JSON.stringify(userData));
                     console.log("userData ", localStorage.getItem('user'));
-                    history.push('/');
+                    navigate('/');
                 })
                 .catch(err => {
                     console.log("currentSession error: ", err);
